requirements/base.txt
gunicorn==19.7.1
marshmallow==3.0.0b7
SQLAlchemy==1.2.5
SQLAlchemy-Utils==0.33.1
coverage==4.5.1
PyJWT==1.6.1
psycopg2==2.7.4
enum34==1.1.6
werkzeug==0.14.1
passlib==1.7.1